+++ src/Game/GameObjects/BrickWall.h
@@ -54,7 +54,7 @@ public:
   BrickWall(const EBrickWallType eBrickWallType, const glm::vec2 &position, const glm::vec2 &size, const float rotation,
     const float layer) noexcept;
   virtual void render() const noexcept override;
-  virtual void update(const uint64_t delta) noexcept override;
+  virtual void update(const double delta) noexcept override;
 private:
   void renderBrick(const EBrickLocation eBrickLocation) const noexcept;
   std::array<EBrickState, 4> m_eCurrentBrickState;
