+++ src/Game/GameObjects/Eagle.h
@@ -19,7 +19,7 @@ public:
   };
   Eagle(const glm::vec2 &position, const glm::vec2 &size, const float rotation, const float layer) noexcept;
   virtual void render() const noexcept override;
-  virtual void update(const uint64_t delta) noexcept override;
+  virtual void update(const double delta) noexcept override;
 private:
   std::array<std::shared_ptr<RendererEngine::Sprite>, 2> m_sprites;
   EEagleState m_eCurrentEagleState;
