+++ src/Renderer/Sprite.h
@@ -24,10 +24,10 @@ namespace RendererEngine
 
     struct FrameDescription
     {
-      FrameDescription(const glm::vec2 _leftBottomUV, const glm::vec2 _rightTopUV, const uint64_t _duration) noexcept;
+      FrameDescription(const glm::vec2 _leftBottomUV, const glm::vec2 _rightTopUV, const double _duration) noexcept;
       glm::vec2 leftBottomUV;
       glm::vec2 rightTopUV;
-      uint64_t duration;
+      double duration;
     };
 
     Sprite(const std::shared_ptr<Texture2D> &pTexture,
@@ -39,7 +39,7 @@ namespace RendererEngine
     void render(const glm::vec2 &position, const glm::vec2 &size, const float rotation, const float layer = 0.f,
       const size_t frameId = 0) const noexcept;
 
-    uint64_t getFrameDuration(const size_t frameId) const noexcept;
+    double getFrameDuration(const size_t frameId) const noexcept;
 
     size_t getFramesCount() const noexcept;
