+++ src/Renderer/SpriteAnimator.h
@@ -9,12 +9,12 @@ namespace RendererEngine {
   public:
     SpriteAnimator(std::shared_ptr<Sprite> pSprite) noexcept;
     size_t getCurrentFrame() const noexcept;
-    void update(const uint64_t delta) noexcept;
+    void update(const double delta) noexcept;
   private:
     std::shared_ptr<Sprite> m_pSprite;
     size_t m_currentFrame;
-    uint64_t m_currentFrameDuration;
-    uint64_t m_currentAnimationTime;
+    double m_currentFrameDuration;
+    double m_currentAnimationTime;
   };
 }
 #endif // !_SPRITEANIMATOR_H
